Collapse all models into one large class
The problem with using subtypes is I loose the binding infromation when I store them all in a collection.
I can recreate the binding when I typecast them but this doesn't work with primative types (bools).
SwiftUI appears to lose the binding between the model and view object.
I don't really understand what's happening, but if keep the bindings intact it all works